Use setuptools to install requirements
Current setup was using Python's distutils which does not allow setting
any required modules for the package. Change the flow of setup making it
use setuptools whenever possible to be able to install the requirements,
fallback to distutils otherwise. Warning about the option not recognized
by distutils has also been taken care of with this patch in case the
installation is done with distutils.
Now, suricata-update installs as a complete application taking care of
all the requirements.

Closes redmine ticket #2667