Cleanup main import 'import types'
Optimization Cleanup main imports

Using isinstance() is the preferred way to access the type instead
of using the types module. Can be done by removing 'import types'
and using isinstance(). The isinstance() function checks if the object
(first argument) is an instance or subclass of classinfo class
(second argument). Here object is the 'url' to be checked and classinfo
is a string type.