Improve permission warning
Improve permission warning when Suricata-update runs with the wrong user

When `suricata-update` runs with a non-root user, it gives an ugly traceback.
To avoid those ugly tracebacks, the permission of the data-directory is
checked where the file `suricata.yaml` is loaded from and exit cleanly with
an error in the log.
The output now looks like:

```
11/4/2019 -- 01:14:42 - <Info> -- Using data-directory /usr/local/var/lib/suricata.
11/4/2019 -- 01:14:42 - <Info> -- Using Suricata configuration /usr/local/etc/suricata/suricata.yaml
11/4/2019 -- 01:14:42 - <Info> -- Using /usr/local/etc/suricata/rules for Suricata provided rules.
11/4/2019 -- 01:14:42 - <Info> -- Found Suricata version 5.0.0-dev at /usr/local/bin/suricata.
11/4/2019 -- 01:14:42 - <Info> -- Loading /usr/local/etc/suricata/suricata.yaml
11/4/2019 -- 01:14:42 - <Error> -- [ERRCODE: SC_ERR_FATAL(171)] - failed to open file: /usr/local/etc/suricata/suricata.yaml: Permission denied
11/4/2019 -- 01:14:42 - <Error> -- Permission denied: Please alter the permissions for: /usr/local/var/lib/suricata
```