Improve permission warning
Improve permission warning when Suricata-update runs with the wrong user

When suricata-update runs with a non-root user, it gives an ugly traceback.
To avoid those ugly tracebacks, the permission of the configuration file
and rule file are checked where the files are loaded from configuration
file directory and rule output directory respectively and exit cleanly
with an error in the log.