Improve permission warning
Improve permission warning when Suricata-update runs with the wrong user

When suricata-update runs with a non-root user, it gives an ugly traceback.
To avoid those ugly tracebacks, the permission of the file is checked earlier
and try except block is put around the operations that are being performed on
the file and exit cleanly with an error in the log.