Improve permission warnings
Improve permission warning when Suricata-update runs with the wrong user

When suricata-update runs with a non-root user, it gives an ugly traceback.
To avoid those ugly tracebacks, try except block is put around the operation
that are being performed on the file i.e. `shutil.copyfile(src, dest)` and
exit cleanly with an error in the log.
Also, to avoid ugly tracebacks for permission of `suricata.yaml`, try-except
is put around where it is accessed and exit cleanly with an error in the log.